# Add new intrinsic `Paths` to avoid snapshotting