--- a/src/python/pants/engine/internals/graph.py
+++ b/src/python/pants/engine/internals/graph.py
@@ -30,6 +30,7 @@ from pants.engine.collection import Collection
 from pants.engine.fs import (
     EMPTY_SNAPSHOT,
     Digest,
+    FileListing,
     MergeDigests,
     PathGlobs,
     Snapshot,
@@ -102,15 +103,15 @@ async def generate_subtargets(address: Address, global_options: GlobalOptions) -
         return Subtargets(base_target, ())
 
     # Generate a subtarget per source.
-    snapshot = await Get(Snapshot, PathGlobs, sources_field_path_globs)
-    sources_field.validate_snapshot(snapshot)
+    file_listing = await Get(FileListing, PathGlobs, sources_field_path_globs)
+    sources_field.validate_resolved_files(file_listing.files)
     wrapped_subtargets = await MultiGet(
         Get(
             WrappedTarget,
             Address,
             generate_subtarget_address(address, full_file_name=subtarget_file),
         )
-        for subtarget_file in snapshot.files
+        for subtarget_file in file_listing.files
     )
 
     return Subtargets(base_target, tuple(wt.target for wt in wrapped_subtargets))
@@ -339,9 +340,9 @@ class Owners(Collection[Address]):
 @rule
 async def find_owners(owners_request: OwnersRequest) -> Owners:
     # Determine which of the sources are live and which are deleted.
-    sources_set_snapshot = await Get(Snapshot, PathGlobs(owners_request.sources))
+    sources_set_file_listing = await Get(FileListing, PathGlobs(owners_request.sources))
 
-    live_files = FrozenOrderedSet(sources_set_snapshot.files)
+    live_files = FrozenOrderedSet(sources_set_file_listing.files)
     deleted_files = FrozenOrderedSet(s for s in owners_request.sources if s not in live_files)
     live_dirs = FrozenOrderedSet(os.path.dirname(s) for s in live_files)
     deleted_dirs = FrozenOrderedSet(os.path.dirname(s) for s in deleted_files)
@@ -434,9 +435,9 @@ async def addresses_with_origins_from_filesystem_specs(
     exactly one file in its `sources` field.
     """
     owners_not_found_behavior = global_options.options.owners_not_found_behavior
-    snapshot_per_include = await MultiGet(
+    file_listing_per_include = await MultiGet(
         Get(
-            Snapshot,
+            FileListing,
             PathGlobs,
             filesystem_specs.path_globs_for_spec(
                 spec, owners_not_found_behavior.to_glob_match_error_behavior()
@@ -445,7 +446,8 @@ async def addresses_with_origins_from_filesystem_specs(
         for spec in filesystem_specs.includes
     )
     owners_per_include = await MultiGet(
-        Get(Owners, OwnersRequest(sources=snapshot.files)) for snapshot in snapshot_per_include
+        Get(Owners, OwnersRequest(sources=file_listing.files))
+        for file_listing in file_listing_per_include
     )
     addresses_to_specs: Dict[Address, FilesystemSpec] = {}
     for spec, owners in zip(filesystem_specs.includes, owners_per_include):
@@ -629,7 +631,7 @@ async def hydrate_sources(
     if path_globs is None:
         return HydratedSources(EMPTY_SNAPSHOT, sources_field.filespec, sources_type=sources_type)
     snapshot = await Get(Snapshot, PathGlobs, path_globs)
-    sources_field.validate_snapshot(snapshot)
+    sources_field.validate_resolved_files(snapshot.files)
 
     # Finally, return if codegen is not in use; otherwise, run the relevant code generator.
     if not request.enable_codegen or generate_request_type is None: