--- a/src/python/pants/engine/internals/scheduler.py
+++ b/src/python/pants/engine/internals/scheduler.py
@@ -24,10 +24,10 @@ from pants.engine.fs import (
     DigestSubset,
     DownloadFile,
     FileContent,
-    FileListing,
     MergeDigests,
     PathGlobs,
     PathGlobsAndRoot,
+    Paths,
     RemovePrefix,
     Snapshot,
 )
@@ -131,7 +131,7 @@ class Scheduler:
         types = PyTypes(
             directory_digest=Digest,
             snapshot=Snapshot,
-            file_listing=FileListing,
+            paths=Paths,
             file_content=FileContent,
             digest_contents=DigestContents,
             address=Address,