--- a/src/python/pants/source/source_root.py
+++ b/src/python/pants/source/source_root.py
@@ -10,7 +10,7 @@ from typing import Dict, Iterable, Optional, Set, Tuple, Union
 
 from pants.build_graph.address import Address
 from pants.engine.collection import DeduplicatedCollection
-from pants.engine.fs import PathGlobs, Snapshot
+from pants.engine.fs import FileListing, PathGlobs
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import Target
 from pants.option.subsystem import Subsystem
@@ -279,9 +279,10 @@ async def get_optional_source_root(
                 raise InvalidMarkerFileError(
                     f"Marker filename must be a base name: {marker_filename}"
                 )
-        # TODO: An intrinsic to check file existence at a fixed path?
-        snapshot = await Get(Snapshot, PathGlobs([str(path / mf) for mf in marker_filenames]))
-        if len(snapshot.files) > 0:
+        file_listing = await Get(
+            FileListing, PathGlobs([str(path / mf) for mf in marker_filenames])
+        )
+        if len(file_listing.files) > 0:
             return OptionalSourceRoot(SourceRoot(str(path)))
 
     # The requested path itself is not a source root, but maybe its parent is.
@@ -329,16 +330,16 @@ async def all_roots(source_root_config: SourceRootConfig) -> AllSourceRoots:
         marker_file_matches.add(f"**/{marker_filename}")
 
     # Match the patterns against actual files, to find the roots that actually exist.
-    pattern_snapshot, marker_file_snapshot = await MultiGet(
-        Get(Snapshot, PathGlobs(globs=sorted(pattern_matches))),
-        Get(Snapshot, PathGlobs(globs=sorted(marker_file_matches))),
+    pattern_file_listing, marker_file_listing = await MultiGet(
+        Get(FileListing, PathGlobs(globs=sorted(pattern_matches))),
+        Get(FileListing, PathGlobs(globs=sorted(marker_file_matches))),
     )
 
     responses = await MultiGet(
         itertools.chain(
             (
                 Get(OptionalSourceRoot, SourceRootRequest(PurePath(d)))
-                for d in pattern_snapshot.dirs
+                for d in pattern_file_listing.dirs
             ),
             # We don't technically need to issue a SourceRootRequest for the marker files,
             # since we know that their immediately enclosing dir is a source root by definition.
@@ -346,7 +347,7 @@ async def all_roots(source_root_config: SourceRootConfig) -> AllSourceRoots:
             # logic here.
             (
                 Get(OptionalSourceRoot, SourceRootRequest(PurePath(f)))
-                for f in marker_file_snapshot.files
+                for f in marker_file_listing.files
             ),
         )
     )