--- a/src/rust/engine/fs/src/lib.rs
+++ b/src/rust/engine/fs/src/lib.rs
@@ -203,17 +203,11 @@ pub enum PathStat {
 
 impl PathStat {
   pub fn dir(path: PathBuf, stat: Dir) -> PathStat {
-    PathStat::Dir {
-      path: path,
-      stat: stat,
-    }
+    PathStat::Dir { path, stat }
   }
 
   pub fn file(path: PathBuf, stat: File) -> PathStat {
-    PathStat::File {
-      path: path,
-      stat: stat,
-    }
+    PathStat::File { path, stat }
   }
 
   pub fn path(&self) -> &Path {
@@ -221,6 +215,31 @@ impl PathStat {
       &PathStat::Dir { ref path, .. } => path.as_path(),
       &PathStat::File { ref path, .. } => path.as_path(),
     }
+  }
+
+  ///
+  /// Sort and ensure that there were no duplicate entries.
+  ///
+  /// TODO: audit if this is even necessary? expand_globs() already sorts and dedupes, but, are
+  /// there other callers of this where we can't be confident the sorting happened? Maybe, we
+  /// should add a light wrapper around `Vec<PathStat>` that ensures it's sorted and deduped, as
+  /// we're now using the type a lot.
+  pub fn normalize_path_stats(mut path_stats: Vec<PathStat>) -> Result<Vec<PathStat>, String> {
+    #[allow(clippy::unnecessary_sort_by)]
+    path_stats.sort_by(|a, b| a.path().cmp(b.path()));
+
+    // The helper assumes that if a Path has multiple children, it must be a directory.
+    // Proactively error if we run into identically named files, because otherwise we will treat
+    // them like empty directories.
+    let pre_dedupe_len = path_stats.len();
+    path_stats.dedup_by(|a, b| a.path() == b.path());
+    if path_stats.len() != pre_dedupe_len {
+      return Err(format!(
+        "Snapshots must be constructed from unique path stats; got duplicates in {:?}",
+        path_stats
+      ));
+    }
+    Ok(path_stats)
   }
 }
 
@@ -693,7 +712,7 @@ impl PathStatGetter<io::Error> for Arc<PosixFS> {
               async move {
                 match maybe_stat {
                   // Note: This will drop PathStats for symlinks which don't point anywhere.
-                  Some(Stat::Link(link)) => fs.canonicalize(link.0.clone(), link).await,
+                  Some(Stat::Link(link)) => fs.canonicalize_link(link.0.clone(), link).await,
                   Some(Stat::Dir(dir)) => Ok(Some(PathStat::dir(dir.0.clone(), dir))),
                   Some(Stat::File(file)) => Ok(Some(PathStat::file(file.path.clone(), file))),
                   None => Ok(None),