--- a/src/rust/engine/fs/src/posixfs_tests.rs
+++ b/src/rust/engine/fs/src/posixfs_tests.rs
@@ -357,7 +357,7 @@ async fn memfs_expand_basic() {
   .unwrap();
 
   assert_eq!(
-    fs.expand(globs).await.unwrap(),
+    fs.expand_globs(globs).await.unwrap(),
     vec![
       PathStat::file(
         p1.clone(),