### Problem

Several of our mission-critical rules only need to resolve file names to work; they do not need the actual file contents, nor to save the snapshot to LMDB store. This includes:

- Resolving `AddressInput -> Address`.
- Getting the directory names for `AddressSpecs` for which we should get `AddressFamily`.
- Generating subtargets to figure out which subtargets exist.
- Finding owners, which requires determining what is a real file vs. deleted.
- Determining source roots via marker files.

### Solution

Add a new intrinsic `Paths`, which is nearly identical to `Snapshot`, but leaves off the `digest` field and does not snapshot the directory.

### Result

This command will run several of the above rules, including finding owners, resolving address specs, and generating subtargets.

Before:

```
multitime -n 50 ./pants --no-pantsd --changed-since=release_2.0.0a0 list
            Mean        Std.Dev.    Min         Median      Max
real        2.848       0.080       2.797       2.823       3.279
user        2.905       0.043       2.851       2.890       3.057
sys         1.757       0.023       1.721       1.752       1.826

```

After:

```
multitime -n 50 ./pants --no-pantsd --changed-since=release_2.0.0a0 list
            Mean        Std.Dev.    Min         Median      Max
real        2.713       0.048       2.628       2.703       2.991
user        2.482       0.024       2.444       2.477       2.552
sys         1.377       0.017       1.327       1.379       1.405

```

This results in a ~5% speedup for real time, ~15% speedup for user time, and ~20% speedup for sys time (ignoring standard deviations).

Closes https://github.com/pantsbuild/pants/issues/10711.

[ci skip-rust]
[ci skip-build-wheels]